Movendo tipo de horario disponivel
