+++ src/TechMedicos/Core/TechMedicos.Domain/ValueObjects/AgendamentoMedico.cs
@@ -48,22 +48,4 @@ namespace TechMedicos.Domain.ValueObjects
             yield return Horarios;
         }
     }
-
-    public class HorarioDisponivel : ValueObject
-    {
-        public HorarioDisponivel(TimeOnly horaInicio)
-        {
-            HoraInicio = horaInicio;
-            HoraFim = horaInicio.AddMinutes(50);
-        }
-
-        public TimeOnly HoraInicio { get; private set; }
-        public TimeOnly HoraFim { get; private set; }
-
-        protected override IEnumerable<object> RetornarPropriedadesDeEquidade()
-        {
-            yield return HoraInicio;
-            yield return HoraFim;
-        }
-    }
 }
